Set default item props & renamed status component
